LayoutTests/fast/events/touch/ios/prevent-default-with-window-tap-gesture-expected.txt

This test verifies that a gesture recognizer installed on the window containing the web view is not prevented from recognizing if the web page prevents default on touchstart. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Tap gesture on window recognized
PASS receivedTouchStart is true
PASS successfullyParsed is true

TEST COMPLETE
